+++ buildpacks/dotnet/tests/dotnet_publish_test.rs
@@ -68,6 +68,11 @@ fn test_dotnet_publish_with_debug_configuration() {
 
             let rid = get_rid();
             assert_contains!(
+                &context.pack_stdout,
+                &formatdoc! {r#"
+                    - Running `dotnet publish /workspace/foo.csproj --runtime {rid} "-p:PublishDir=bin/publish" --artifacts-path /tmp/build_artifacts --configuration Debug`"#}
+            );
+            assert_contains!(
                 replace_msbuild_log_patterns_with_placeholder(
                     &context.pack_stdout,
                     "<PLACEHOLDER>"
@@ -155,7 +160,6 @@ fn test_dotnet_publish_with_global_json_and_custom_verbosity_level() {
               replace_msbuild_log_patterns_with_placeholder(&context.pack_stdout, "<PLACEHOLDER>"), 
               &formatdoc! {r#"
                 - Publish solution
-                  - Using `Release` build configuration
                   - Running `dotnet publish /workspace/foo.csproj --runtime {rid} "-p:PublishDir=bin/publish" --artifacts-path /tmp/build_artifacts --verbosity normal`
                 
                       MSBuild version 17.8.3+195e7f5a3 for .NET